refactor: BaseEntity 리팩터링
