+++ src/main/java/com/nextdaydelivery/delivery/domain/entity/Delivery.java
@@ -1,7 +1,7 @@
 package com.nextdaydelivery.delivery.domain.entity;
 
 import com.nextdaydelivery.delivery.domain.enums.DeliveryStatus;
-import com.nextdaydelivery.global.baseEntity.BaseEntity;
+import com.nextdaydelivery.global.domain.BaseAuditEntity;
 import com.nextdaydelivery.order.domain.entity.Order;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -13,6 +13,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
+import java.time.LocalDateTime;
 import java.util.UUID;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -27,7 +28,7 @@ import org.hibernate.annotations.GenericGenerator;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class Delivery extends BaseEntity {
+public class Delivery extends BaseAuditEntity {
 
     @Id
     @GeneratedValue(generator = "UUID")
@@ -42,5 +43,10 @@ public class Delivery extends BaseEntity {
     @Enumerated(EnumType.STRING)
     @Column(name = "delivery_status", nullable = false)
     private DeliveryStatus deliveryStatus; // 배달 상태 (PENDING, ING, COMPLETED)
+    
+    @Column(name = "deleted_at")
+    private LocalDateTime deletedAt; // 레코드 삭제 시간
 
+    @Column(name = "deleted_by", length = 100)
+    private String deletedBy; // 레코드 삭제자
 }
