+++ src/main/java/com/nextdaydelivery/payment/domain/enums/PaymentMethod.java
@@ -1,4 +1,4 @@
-package com.nextdaydelivery.payment.domain.entity.enums;
+package com.nextdaydelivery.payment.domain.enums;
 
 /**
  * 결제 방법을 위한 ENUM
