+++ src/main/java/com/nextdaydelivery/payment/domain/enums/PaymentStatus.java
@@ -1,4 +1,4 @@
-package com.nextdaydelivery.payment.domain.entity.enums;
+package com.nextdaydelivery.payment.domain.enums;
 
 /**
  * 결제 상태를 위한 ENUM
