+++ src/main/java/com/nextdaydelivery/product/domain/entity/Product.java
@@ -1,6 +1,6 @@
 package com.nextdaydelivery.product.domain.entity;
 
-import com.nextdaydelivery.global.baseEntity.BaseEntity;
+import com.nextdaydelivery.global.domain.BaseAuditEntity;
 import com.nextdaydelivery.store.domain.entity.Store;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -10,6 +10,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import java.time.LocalDateTime;
 import java.util.UUID;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -24,7 +25,7 @@ import org.hibernate.annotations.GenericGenerator;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class Product extends BaseEntity {
+public class Product extends BaseAuditEntity {
 
     @Id
     @GeneratedValue(generator = "UUID")
@@ -47,4 +48,10 @@ public class Product extends BaseEntity {
 
     @Column(name = "is_hidden", nullable = false)
     private Boolean isHidden; // 숨김 여부 (BOOLEAN)
+    
+    @Column(name = "deleted_at")
+    private LocalDateTime deletedAt; // 레코드 삭제 시간
+
+    @Column(name = "deleted_by", length = 100)
+    private String deletedBy; // 레코드 삭제자
 }
